scripts/pyAMI_atlas/pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = 'pyAMI_atlas'
version = "5.1"
authors = [
    {name = 'Jerome Odier'}
]
description = 'Python ATLAS Metadata Interface (pyAMI) for ATLAS'
dependencies = ['pyAMI-core']
